feature : 상품 찜하기 기능 및 예외 처리 추가

+++ src/main/java/com/softsquared/template/src/user/UserInfoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository // => JPA => Hibernate => ORM => Database 객체지향으로 접근하게 해주는 도구이다
-public interface UserInfoRepository extends CrudRepository<UserInfo, Integer> {
+public interface UserInfoRepository extends CrudRepository<UserInfo, Long> {
 //    List<UserInfo> findByStatus(String status);
 //    List<UserInfo> findByEmailAndStatus(String email, String status);
 //    List<UserInfo> findByStatusAndNicknameIsContaining(String status, String word);
